More refactor Add placeholder for 2 messaging methods

+++ InstaSharper/Classes/Models/Direct/InstaInboxMedia.cs
@@ -7,8 +7,8 @@ namespace InstaSharper.Classes.Models.Direct
     public class InstaInboxMedia
     {
         public List<InstaImage> Images { get; set; } = new List<InstaImage>();
-        public long OriginalWidth { get; set; }
-        public long OriginalHeight { get; set; }
+        public int OriginalWidth { get; set; }
+        public int OriginalHeight { get; set; }
         public InstaMediaType MediaType { get; set; }
         public List<InstaVideo> Videos { get; set; } = new List<InstaVideo>();
     }
