+++ InstaSharper/Classes/ResponseWrappers/Media/InstaInboxMediaResponse.cs
@@ -8,9 +8,9 @@ namespace InstaSharper.Classes.ResponseWrappers.Media
     {
         [JsonProperty("image_versions2")] public InstaImageCandidatesResponse ImageCandidates { get; set; }
 
-        [JsonProperty("original_width")] public long OriginalWidth { get; set; }
+        [JsonProperty("original_width")] public int OriginalWidth { get; set; }
 
-        [JsonProperty("original_height")] public long OriginalHeight { get; set; }
+        [JsonProperty("original_height")] public int OriginalHeight { get; set; }
 
         [JsonProperty("media_type")] public InstaMediaType MediaType { get; set; }
